Remove unused evaluateFormula helper
